Move yaml from utils to io.misc
